Update and organize code
