+++ backend/apps/github/api/internal/dataloaders/repository.py
@@ -7,7 +7,6 @@ from apps.github.models.release import Release
 from apps.github.models.repository import Repository
 
 REPOSITORY_BY_RELEASE_ID_LOADER = "repository_by_release_id"
-RELEASE_URL_BY_ID_LOADER = "release_url_by_id"
 
 
 async def load_repositories_by_release_id(
@@ -23,32 +22,10 @@ async def load_repositories_by_release_id(
     )
 
 
-async def load_release_urls_by_id(release_ids: list[int]) -> list[str]:
-    """Batch-load release URLs for the given release IDs in a single query."""
-    releases = Release.objects.filter(pk__in=release_ids).select_related(
-        "repository__owner", "repository__organization"
-    )
-
-    mapping: dict[int, str] = {}
-    async for release in releases:
-        if release.repository:
-            mapping[release.pk] = f"{release.repository.url}/releases/tag/{release.tag_name}"
-        else:
-            mapping[release.pk] = ""
-
-    return [mapping.get(release_id, "") for release_id in release_ids]
-
-
-def get_repository_loaders() -> dict[
-    str,
-    DataLoader[int, Repository | None] | DataLoader[int, str],
-]:
+def get_repository_loaders() -> dict[str, DataLoader[int, Repository | None]]:
     """Return a mapping of per-request DataLoader instances."""
     return {
         REPOSITORY_BY_RELEASE_ID_LOADER: DataLoader[int, Repository | None](
             load_fn=load_repositories_by_release_id,
-        ),
-        RELEASE_URL_BY_ID_LOADER: DataLoader[int, str](
-            load_fn=load_release_urls_by_id,
         ),
     }
